// opabs-web/src/app/shared/model/NamedCurves.ts
export enum NamedCurves {
  curve25519 = 'curve25519',
  secp128r1 = 'secp128r1',
  secp160k1 = 'secp160k1',
  secp160r1 = 'secp160r1',
  secp160r2 = 'secp160r2',
  secp192k1 = 'secp192k1',
  secp192r1 = 'secp192r1',
  secp224k1 = 'secp224k1',
  secp224r1 = 'secp224r1',
  secp256k1 = 'secp256k1',
  secp256r1 = 'secp256r1',
  secp384r1 = 'secp384r1',
  secp521r1 = 'secp521r1'
}