// opabs-web/src/app/shared/model/KeyUsage.ts
export enum KeyUsage {
  DIGITAL_SIGNATURE = 'Digital Signature',
  NON_REPUDIATION = 'Non Repudiation',
  KEY_ENCIPHERMENT = 'Key Encipherment',
  DATA_ENCIPHERMENT = 'Data Encipherment',
  KEY_AGREEMENT = 'Key Agreement',
  KEY_CERT_SIGN = 'Certificate Signature',
  CRL_SIGN = 'CRL Sign',
  ENCIPHER_ONLY = 'Encipher Only',
  DECIPHER_ONLY = 'Decipher Only'
}